Master Iterative Approach ver4.4, testing RealImages_ver2
 found an error in the update_subgraphs_3.m (casted integer to uint8)
  need to recalculate results test RealImages2/ver1,ver2

 + test RealImages2/ver3 which uses candidate matches for matching on the lower level

+++ Master/IterativeApproach/Testing/RealImages2/ver1/Methods/TwoLevelGM_ver4.3.3/2levelGM/update_partition/update_subgraphs_3.m
@@ -60,7 +60,7 @@ function [HLG1, HLG2] = update_subgraphs_3(LLG1, LLG2, HLG1, HLG2, ...
       [~,ia] = unique(tmp(ind,1));
       tmp = tmp(ind(ia),:);
 
-      new_U2(uint8(tmp(:,1)), aj) = tmp(:,2);   
+      new_U2(tmp(:,1), aj) = tmp(:,2);   
       
       % Project Vaj into LLG1.V
       PVaj = Aj * Vaj' + repmat(bj,1,size(Vaj,1)); % projection of Vaj_nm nodes
@@ -78,7 +78,7 @@ function [HLG1, HLG2] = update_subgraphs_3(LLG1, LLG2, HLG1, HLG2, ...
       [~,ia] = unique(tmp(ind,1));
       tmp = tmp(ind(ia),:);
 
-      new_U1(uint8(tmp(:,1)), ai) = tmp(:,2);
+      new_U1(tmp(:,1), ai) = tmp(:,2);
       
 %       figure; subplot(1,2,1);
 %
